fix(MainPage.tsx): Change directory component to a separate file and add types
