+++ app/src/Pages/MainPage/MainPage.tsx
@@ -1,14 +1,6 @@
-type FileData = {
-  name: string;
-};
+import Directory from './Components/Directory/Directory';
+import DirectoryData from './Types/DirectoryData';
 
-type DirectoryData = {
-  name: string;
-  subdirs: DirectoryData[];
-  files: FileData[];
-};
-
-const DIR_GAP = 100;
 const DIR_DATA: DirectoryData = {
   name: 'Dir1',
   subdirs: [
@@ -33,22 +25,6 @@ const DIR_DATA: DirectoryData = {
       name: 'File2',
     },
   ],
-};
-
-const Directory = ({ dirData }) => {
-  return (
-    <div className="flex flex-col items-end w-full">
-      <div className="w-full text-2xl">{dirData.name}</div>
-      <div style={{ width: `calc(100% - ${DIR_GAP}px)` }}>
-        {dirData.subdirs.map((subdir: DirectoryData) => (
-          <Directory dirData={subdir} />
-        ))}
-        {dirData.files.map((file: FileData) => (
-          <div className="w-full text-2xl">{file.name}</div>
-        ))}
-      </div>
-    </div>
-  );
 };
 
 const MainPage = () => {
