+++ app/src/Pages/MainPage/Types/DirectoryData.tsx
@@ -1,0 +1,9 @@
+import FileData from './FileData';
+
+type DirectoryData = {
+  name: string;
+  subdirs: DirectoryData[];
+  files: FileData[];
+};
+
+export default DirectoryData;
