+++ app/src/Pages/MainPage/Types/FileData.tsx
@@ -1,0 +1,5 @@
+type FileData = {
+  name: string;
+};
+
+export default FileData;
